+++ CHANGELOG.md
@@ -1,6 +1,13 @@
-## 3.0.0
+## v3.0.0
 New features:
+- Add a signal generator capture device.
 - Optionally write wav header when outputting to file or stdout.
+Changes:
+- Filter pipeline steps take a list of channels to filter instead of a single one.
+
+## v2.0.3
+Bugfixes:
+- MacOS: Fix using Aggregate devices for playback.
 
 ## v2.0.2
 Bugfixes: